refactor: move MixNode to node.py
